+++ src/Test/Integration/fixtures/product_with_description.php
@@ -16,7 +16,7 @@
  * limitations under the License.
  *
  * @author    Shopgate Inc, 804 Congress Ave, Austin, Texas 78701 <interfaces@shopgate.com>
- * @copyright 2019 Shopgate Inc
+ * @copyright Shopgate Inc
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 /** @noinspection PhpUndefinedClassInspection */
@@ -28,51 +28,28 @@ use Magento\Catalog\Model\Product\Attribute\Source\Status;
 use Magento\Catalog\Model\Product\Visibility;
 use Magento\Catalog\Model\ProductFactory;
 use Magento\Catalog\Model\ProductRepository;
-use Magento\Customer\Model\GroupManagement;
 use Magento\TestFramework\Helper\Bootstrap;
 
-/** @var Product $product */
-$factory = Bootstrap::getObjectManager()->get(ProductFactory::class);
-$product = $factory->create();
 /** @var ProductRepository $repository */
+$factory    = Bootstrap::getObjectManager()->get(ProductFactory::class);
+$product    = $factory->create();
 $repository = Bootstrap::getObjectManager()->create(ProductRepository::class);
+
+/** @var Product $product */
+$product = $factory->create();
 $product->setTypeId('simple')
         ->setAttributeSetId(4)
         ->setWebsiteIds([1])
-        ->setName('Simple Product')
+        ->setName('Simple Product Two')
         ->setPrice(1000)
         ->setFinalPrice(1000)
-        ->setSku('simple-test-product')
+        ->setSku('simple-test-product-two')
         ->setVisibility(Visibility::VISIBILITY_BOTH)
         ->setStatus(Status::STATUS_ENABLED)
         ->setWeight(15.5)
         ->setStockData(['use_config_manage_stock' => 1, 'qty' => 100, 'is_qty_decimal' => 0, 'is_in_stock' => 1])
-        ->setData(
-            'tier_price',
-            [
-                [
-                    'website_id'       => 0,
-                    'cust_group'       => GroupManagement::CUST_GROUP_ALL,
-                    'price_qty'        => 2,
-                    'price'            => 8,
-                    'percentage_value' => 0
-                ],
-                [
-                    'website_id'       => 0,
-                    'cust_group'       => 1,
-                    'price_qty'        => 5,
-                    'price'            => 50,
-                    'percentage_value' => 50
-                ],
-                [
-                    'website_id'       => 0,
-                    'cust_group'       => GroupManagement::NOT_LOGGED_IN_ID,
-                    'price_qty'        => 3,
-                    'price'            => 20,
-                    'percentage_value' => 0
-                ],
-            ]
-        );
+        ->setData('description', 'Long Description1')
+        ->setData('short_description', 'Short Description2');
 
 /** @noinspection PhpUnhandledExceptionInspection */
 $repository->save($product);